Activities and controller views are now updated according previous refactored stuff
