+++ app/src/main/kotlin/com/cesarvaliente/kunidirectional/itemslist/ItemsControllerView.kt
@@ -20,45 +20,38 @@
 package com.cesarvaliente.kunidirectional.itemslist
 
 import com.cesarvaliente.kunidirectional.ControllerView
-import com.cesarvaliente.kunidirectional.actionDispatcherSingleton
-import com.cesarvaliente.kunidirectional.stateDispatcherSingleton
-import com.cesarvaliente.kunidirectional.store.ActionDispatcher
+import com.cesarvaliente.kunidirectional.store.DeleteAction.DeleteItemAction
 import com.cesarvaliente.kunidirectional.store.Item
 import com.cesarvaliente.kunidirectional.store.Navigation
+import com.cesarvaliente.kunidirectional.store.NavigationAction.EditItemScreenAction
+import com.cesarvaliente.kunidirectional.store.ReadAction.FetchItemsAction
 import com.cesarvaliente.kunidirectional.store.State
-import com.cesarvaliente.kunidirectional.store.StateDispatcher
+import com.cesarvaliente.kunidirectional.store.Store
 import com.cesarvaliente.kunidirectional.store.ThreadExecutor
-import com.cesarvaliente.kunidirectional.store.action.DeleteAction.DeleteItemAction
-import com.cesarvaliente.kunidirectional.store.action.NavigationAction.EditItemScreenAction
-import com.cesarvaliente.kunidirectional.store.action.ReadAction.FetchItemsAction
-import com.cesarvaliente.kunidirectional.store.action.UpdateAction.ReorderItemsAction
-import com.cesarvaliente.kunidirectional.store.action.UpdateAction.UpdateFavoriteAction
+import com.cesarvaliente.kunidirectional.store.UpdateAction.ReorderItemsAction
+import com.cesarvaliente.kunidirectional.store.UpdateAction.UpdateFavoriteAction
 import java.lang.ref.WeakReference
 
 class ItemsControllerView(
         val itemsViewCallback: WeakReference<ItemsViewCallback>,
-        actionDispatcher: ActionDispatcher = actionDispatcherSingleton,
-        stateDispatcher: StateDispatcher = stateDispatcherSingleton,
-        handleStateDifferentThread: ThreadExecutor? = null)
-    : ControllerView(actionDispatcher, stateDispatcher, handleStateDifferentThread) {
+        store: Store,
+        mainThread: ThreadExecutor? = null)
+    : ControllerView(store, mainThread) {
 
     fun fetchItems() =
-            dispatch(FetchItemsAction())
+            store.dispatch(FetchItemsAction())
 
     fun toEditItemScreen(item: Item) =
-            dispatch(EditItemScreenAction(item))
+            store.dispatch(EditItemScreenAction(item))
 
     fun reorderItems(items: List<Item>) =
-            dispatch(ReorderItemsAction(items))
+            store.dispatch(ReorderItemsAction(items))
 
     fun changeFavoriteStatus(item: Item) =
-            dispatch(UpdateFavoriteAction(
-                    localId = item.localId,
-                    favorite = !item.favorite
-            ))
+            store.dispatch(UpdateFavoriteAction(localId = item.localId, favorite = !item.favorite))
 
     fun deleteItem(item: Item) =
-            dispatch(DeleteItemAction(item.localId))
+            store.dispatch(DeleteItemAction(item.localId))
 
     override fun handleState(state: State) {
         when (state.navigation) {
